chore(nextcloud): wire psalm into composer and CI
psalm.xml was committed but dead: vimeo/psalm was in neither composer.json nor
composer.lock, no CI job ran it, and its schemaLocation pointed at a file that
did not exist. #423 replaced the hand-written OCP stubs with nextcloud/ocp,
which is published for static analysers, so psalm can finally resolve OCP\*.

- add vimeo/psalm to require-dev plus a `composer psalm` script, and run it in
  the nextcloud-app CI job
- point psalm at vendor/ via <extraFiles> so OCP\* resolves, and register the
  private-OC\ stubs; stub OC\Core\Command\Base and the global OC class, which
  nextcloud/ocp does not carry
- suppress the issue classes that a Nextcloud app structurally trips (DI-built
  classes read as unused, Entity magic properties) with the reason recorded in
  psalm.xml; no baseline file

Fixing the ~380 findings surfaced real bugs and doc gaps:

- UsageStatMapper passed an alias as a second argument to IFunctionBuilder::sum(),
  which takes one — the aliases were dropped, so every per-user token total read 0
- controllers passed the nullable injected user id straight into mappers and
  FileService, which take a non-null string; they now go through
  RequiresUserIdTrait::requireUserId()
- controller @return docblocks omitted the error statuses and half the response
  fields they actually return; the regenerated OpenAPI spec now documents them
- unchecked json_encode()/preg_replace()/end()/stream_get_contents() results,
  and TaskProcessing inputs used without a type guard

Closes #426

+++ nextcloud-app/lib/Service/ImageOptimizer.php
@@ -113,9 +113,9 @@ class ImageOptimizer {
         }
 
         try {
-            $scale = self::MAX_LONG_EDGE / max($width, $height);
-            $newWidth = (int) round($width * $scale);
-            $newHeight = (int) round($height * $scale);
+            $scale = (float)self::MAX_LONG_EDGE / (float)max($width, $height);
+            $newWidth = (int) round((float)$width * $scale);
+            $newHeight = (int) round((float)$height * $scale);
 
             $dst = imagescale($src, $newWidth, $newHeight, IMG_BICUBIC);
             if ($dst === false) {
